+++ src/clientversion.cpp
@@ -11,26 +11,16 @@
 
 /**
  * Name of client reported in the 'version' message. Report the same name
- * for both yacoind and yacoin-core, to make it harder for attackers to
+ * for both yacoind and yacoin-qt, to make it harder for attackers to
  * target servers or GUI users specifically.
  */
 const std::string
-    #ifdef QT_GUI
-        #ifdef _MSC_VER
-            CLIENT_NAME("Yacoin-WM Windows Qt");
-        #elif defined(__APPLE__)
-            CLIENT_NAME("Yacoin-WM MacOS Qt");
-        #else
-            CLIENT_NAME("Yacoin-WM Qt");
-        #endif
+    #if defined(_WIN32) || defined(WIN32)
+        CLIENT_NAME("Yacoin-WM Windows");
+    #elif defined(__APPLE__)
+        CLIENT_NAME("Yacoin-WM MacOS");
     #else
-        #ifdef _MSC_VER
-            CLIENT_NAME("Yacoin-WM Windows daemon");
-        #elif defined(__APPLE__)
-            CLIENT_NAME("Yacoin-WM MacOS daemon");
-        #else
-            CLIENT_NAME("Yacoin-WM daemon");
-        #endif
+        CLIENT_NAME("Yacoin-WM");
     #endif
 
 #define DB_SUFFIX "-leveldb"
