+++ src/init.cpp
@@ -62,7 +62,6 @@
 #include <boost/filesystem/fstream.hpp>
 #include <boost/filesystem/convenience.hpp>
 #include <boost/interprocess/sync/file_lock.hpp>
-#include <boost/algorithm/string/predicate.hpp>
 #include <boost/thread.hpp>
 #include <openssl/crypto.h>
 
@@ -172,22 +171,6 @@ public:
 
 static CCoinsViewErrorCatcher *pcoinscatcher = nullptr;
 static std::unique_ptr<ECCVerifyHandle> globalVerifyHandle;
-
-void WaitForShutdown(boost::thread_group* threadGroup)
-{
-    bool fShutdown = ShutdownRequested();
-    // Tell the main threads to shutdown.
-    while (!fShutdown)
-    {
-        MilliSleep(200);
-        fShutdown = ShutdownRequested();
-    }
-    if (threadGroup)
-    {
-        Interrupt(*threadGroup);
-        threadGroup->join_all();
-    }
-}
 
 void Interrupt(boost::thread_group& threadGroup)
 {
@@ -1609,183 +1592,3 @@ bool AppInitMain(boost::thread_group& threadGroup, CScheduler& scheduler)
 
     return !fRequestShutdown;
 }
-
-//////////////////////////////////////////////////////////////////////////////
-//
-// Start
-//
-#if !defined(QT_GUI) && !defined(TESTS_ENABLED)
-bool AppInit(int argc, char* argv[])
-{
-    boost::thread_group threadGroup;
-    CScheduler scheduler;
-
-    bool fRet = false;
-
-    //
-    // Parameters
-    //
-    // If Qt is used, parameters/bitcoin.conf are parsed in qt/bitcoin.cpp's main()
-    gArgs.ParseParameters(argc, argv);
-
-    // Process help and version before taking care about datadir
-    if (gArgs.IsArgSet("-?") || gArgs.IsArgSet("-h") ||  gArgs.IsArgSet("-help") || gArgs.IsArgSet("-version") || gArgs.IsArgSet("-v"))
-    {
-        std::string strUsage = "Yacoin version: " + FormatFullVersion() + " " + CLIENT_DATE + "\n";
-
-        if (gArgs.IsArgSet("-version") || gArgs.IsArgSet("-v"))
-        {
-            // Do nothing
-//            strUsage += FormatParagraph(LicenseInfo());
-        }
-        else
-        {
-            strUsage += "\n" + _("Usage:") + "\n" +
-                    "  yacoind [options]                     " + "\n" +
-                    "  yacoind [options] <command> [params]  " + _("Send command to -server or yacoind") + "\n" +
-                    "  yacoind [options] help                " + _("List commands") + "\n" +
-                    "  yacoind [options] help <command>      " + _("Get help for a command") + "\n";
-
-            strUsage += "\n" + HelpMessage(HMM_BITCOIND);
-        }
-
-        fprintf(stdout, "%s", strUsage.c_str());
-        return true;
-    }
-
-    try
-    {
-        if (!fs::is_directory(GetDataDir(false)))
-        {
-            fprintf(stderr, "Error: Specified data directory \"%s\" does not exist.\n", gArgs.GetArg("-datadir", "").c_str());
-            return false;
-        }
-
-        try
-        {
-            gArgs.ReadConfigFile(gArgs.GetArg("-conf", YACOIN_CONF_FILENAME));
-        } catch (const std::exception& e) {
-            fprintf(stderr,"Error reading configuration file: %s\n", e.what());
-            return false;
-        }
-
-        // Check for -testnet or -regtest parameter (Params() calls are only valid after this clause)
-        try {
-            SelectParams(ChainNameFromCommandLine());
-        } catch (const std::exception& e) {
-            fprintf(stderr, "Error: %s\n", e.what());
-            return false;
-        }
-
-        bool fCommandLine = false;
-        // Command-line RPC
-        for (int i = 1; i < argc; ++i)
-        {
-            if (!IsSwitchChar(argv[i][0]) && !boost::algorithm::istarts_with(argv[i], "yacoin:"))
-            {
-                fCommandLine = true;
-            }
-        }
-
-        if (fCommandLine)
-        {
-            int ret = CommandLineRPC(argc, argv);
-            exit(ret);
-        }
-        else {
-            // -server defaults to true for bitcoind but not for the GUI so do this here
-            gArgs.SoftSetBoolArg("-server", true);
-            // Set this early so that parameter interactions go to console
-            InitLogging();
-            InitParameterInteraction();
-            if (!AppInitBasicSetup())
-            {
-                // InitError will have been called with detailed error, which ends up on console
-                exit(EXIT_FAILURE);
-            }
-            if (!AppInitParameterInteraction())
-            {
-                // InitError will have been called with detailed error, which ends up on console
-                exit(EXIT_FAILURE);
-            }
-            if (!AppInitSanityChecks())
-            {
-                // InitError will have been called with detailed error, which ends up on console
-                exit(EXIT_FAILURE);
-            }
-//#if !defined(WIN32) && !defined(QT_GUI)
-//            if (fDaemon)
-//            {
-//                // Daemonize
-//                pid_t pid = fork();
-//                if (pid < 0)
-//                {
-//                    fprintf(stderr, "Error: fork() returned %d errno %d\n", pid, errno);
-//                    return false;
-//                }
-//                if (pid > 0)
-//                {
-//                    CreatePidFile(GetPidFile(), pid);
-//                    return true;
-//                }
-//
-//                pid_t sid = setsid();
-//                if (sid < 0)
-//                    fprintf(stderr, "Error: setsid() returned %d errno %d\n", sid, errno);
-//            }
-//#endif
-            if (gArgs.GetBoolArg("-daemon", false))
-            {
-#if HAVE_DECL_DAEMON
-                fprintf(stdout, "Yacoin server starting\n");
-
-                // Daemonize
-                if (daemon(1, 0)) { // don't chdir (1), do close FDs (0)
-                    fprintf(stderr, "Error: daemon() failed: %s\n", strerror(errno));
-                    return false;
-                }
-#else
-                fprintf(stderr, "Error: -daemon is not supported on this operating system\n");
-                return false;
-#endif // HAVE_DECL_DAEMON
-            }
-            // Lock data directory after daemonization
-            if (!AppInitLockDataDirectory())
-            {
-                // If locking the data directory failed, exit immediately
-                exit(EXIT_FAILURE);
-            }
-
-            fRet = AppInitMain(threadGroup, scheduler);
-        }
-    }
-    catch (const std::exception& e) {
-        PrintExceptionContinue(&e, "AppInit()");
-    } catch (...) {
-        PrintExceptionContinue(nullptr, "AppInit()");
-    }
-    if (!fRet)
-    {
-        Interrupt(threadGroup);
-        threadGroup.join_all();
-    }
-    else
-    {
-        WaitForShutdown(&threadGroup);
-    }
-    Shutdown();
-    return fRet;
-}
-
-extern void noui_connect();
-int main(int argc, char* argv[])
-{
-    SetupEnvironment();
-    nUpTimeStart = GetTime();
-
-    // Connect bitcoind signal handlers
-    noui_connect();
-
-    return (AppInit(argc, argv) ? EXIT_SUCCESS : EXIT_FAILURE);
-}
-#endif
